+++ decidim-meetings/app/assets/config/admin/decidim_meetings_manifest.js
@@ -1,3 +1,3 @@
 //= link decidim/meetings/admin/meetings_form.js
 //= link decidim/meetings/admin/registrations_form.js
-//= link decidim/meetings/admin/meetings_form.js
+//= link decidim/meetings/admin/questionnaires.js