+++ decidim-meetings/app/controllers/decidim/meetings/registrations_controller.rb
@@ -4,6 +4,36 @@ module Decidim
   module Meetings
     # Exposes the registration resource so users can join and leave meetings.
     class RegistrationsController < Decidim::Meetings::ApplicationController
+      include FormFactory
+
+      helper_method :meeting, :questionnaire
+
+      def show
+        enforce_permission_to :join, :meeting, meeting: meeting
+
+        return redirect_to meeting_path(meeting) unless questionnaire
+
+        @form = form(QuestionnaireForm).from_model(questionnaire)
+      end
+
+      def answer
+        enforce_permission_to :join, :meeting, meeting: meeting
+
+        @form = form(QuestionnaireForm).from_params(params)
+
+        JoinMeeting.call(meeting, current_user, @form) do
+          on(:ok) do
+            flash[:notice] = I18n.t("registrations.create.success", scope: "decidim.meetings")
+            redirect_to meeting_path(meeting)
+          end
+
+          on(:invalid) do
+            flash.now[:alert] = I18n.t("registrations.create.invalid", scope: "decidim.meetings")
+            render :show
+          end
+        end
+      end
+
       def create
         enforce_permission_to :join, :meeting, meeting: meeting
 
@@ -42,10 +72,18 @@ module Decidim
         @meeting ||= Meeting.where(component: current_component).find(params[:meeting_id])
       end
 
+      def questionnaire
+        @questionnaire ||= Questionnaire.includes(questions: :answer_options).find_by(meeting: meeting.id, questionnaire_type: "registration")
+      end
+
       def redirect_after_path
         referer = request.headers["Referer"]
         return redirect_to(meeting_path(meeting)) if referer =~ /invitation_token/
         redirect_back fallback_location: meeting_path(meeting)
+      end
+
+      def user_has_no_permission_path
+        meeting_path(meeting)
       end
     end
   end