+++ decidim-meetings/db/migrate/20180524112035_add_registration_form_to_meetings.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddRegistrationFormToMeetings < ActiveRecord::Migration[5.1]
+  def change
+    add_column :decidim_meetings_meetings, :registration_form_id, :integer
+  end
+end